Freeboard Dashboard Files
README.md contains only Freeboard info. should perhaps not replace readme for LJordan. was created that way in case of a separate repo for freeboard.
db_copy.sqlite3 is a test db
config.yaml file was changed during testing to reflect that db and is the data show on the included DashboardExample.jpg